+++ routes/auth.js
@@ -1,7 +1,7 @@
 
 const router = require("express").Router();
 
-const authController = require("../controllers/authController");
+const authController = require("../controllers/authConroller/authController");
 
 // auth route
 router.post("/login", [], (req, res) => {
